requirements.txt
boto3==1.23.6
glob2==0.7
psycopg2_binary==2.9.3
python_dateutil==2.8.2
selenium==4.1.5
watchdog==2.1.8
dataclasses_json==0.5.7